Add sloped random generation
Math.random() is linear with each random value having the same possibility of
occuring. Adding Math.sin() to the logic skews the possible values to the larger
side so that they are now a curve insetad of linear so that bigger values are
more prevalent.

Removed min === max check because it makes rollName mocking weird if it
sometimes calls the roll and sometimes does not.